requirements.txt
pre-commit 
detect-secrets
requests
openpyxl
python-dotenv